Change `ensure => present` to `ensure => file` for file resources
While `ensure => present` is perfectly valid, best practices dictate
we not use it, as Linux sees both files and directories as files on
a block device. Consequently, `ensure => file` is more appropriate.

Additionally, we want to get users in the habit of doing this in
their code, as to not cause potential errors or security risks in
the future.

Train early, train well.