Change `mode` to `0644` and `0755` to `0664` and `0775`, respectively
Due to the behavior in Microsoft Windows, we often suggest people
set the group write bit as a default behavior. It is often the case
that Puppet permissions in Windows are granted via group access, so
it is easy to get in a situation where Puppet will write out a file
that it then can no longer access.

Additionally, in the examples used here, the owner and group are
the same, so setting `mode` to `0664` and `0775` have not pratical
effect on the outcome, but gets the learner in the habit.