Handle variation for trailing slash on directory
Removed trailing slash from `$doc_root` variable instantiation.
Updated the variables use throughout the rest of the class, as well
as the underlying tests and harness. Test was updated to handle
either variation (with or without trailing slash).

In general, it is usually better to leave the slash off a variable
declaration and put it where the variable is actually used. While
this does slightly increase character count, it makes a difference
when it is being in conjunction with parameters. By doing it this
way, whether a user passes in a directory with or without a
trailing slash, it will still work on the underlying operating
system, as it will collapse '//' to a '/'.