Update mysql.md account security  task
The account security task is a bit ambiguously stated:

"Go back to your site.pp manifest and include the mysql::server::account_security class in the learning.puppetlabs.vm node. Remember, you don't need to pass any parameters to this class, so a simple include statement will work ***in place of a parameterized class declaration.***"

My initial read of this was that the ::mysql::server::account_security would auto-require the ::mysql::server class, so I should  replace the class definition with the include directive.  This of course leads to an error.

Adding the suggested phrasing would make it clear that the "include ::mysql::server::account_security" should come after the ::mysql::server class definition.