Edit review of the setup steps
I removed the step about MD5 sum because ... if they are looking at this, they successfully unzipped the download, right? The MD5 instructions are on the download page already, and it made for a very unfriendly first step to document "Your download maybe failed, here's a utility to run." If you need to add it back in, maybe add it at the point where they'll fail (when they try to import the OVA?) as a Note, rather than a step.

The step that starts with "Start the VM" is very long, and seems to contain many steps. Are they in the right order? Do we want people running ssh at this point? Or do we just want them to have installed iTerm or PuTTY? If they shouldn't be running ssh right NOW, while the VM is starting, then delete the sentence that tells them to do it. Or move it till after the "It can take several minutes" sentence? I'm not sure.