Edits for README
You use both "reader" and "user." I would choose one of these terms and be consistent. My assumption is that if someone is reading, they are also using the VM. Perhaps that is not a correct assumption and these really are different groups.
Line 35: I am not convinced this is necessary in a section about the user, but I left it in: "Content related to the Puppet
agent/master architecture, Puppet code, and module structure is generally
applicable to the open source version of Puppet, though there are some
differences in file locations."
Line 37: I changed this to agent/master based on architecture docs. Not sure if there is a specific reason to not change it.