Edits for agent run
Line 46: I was not sure what word was missing, but something needs to be adjusted there.

Line 110: I think the tasks get lost in the certificate section. Should they have heading formats?

Line 296: Have you defined or introduced namevar before this quest?

Line 277-296: This could be just my lack of understanding, but you are switching between 'notify' and 'message' as if they are the same. Is 'message' a type of 'notify' resource? The usage of these terms was unclear to me in this section.