(LEARNVM-628) quest setup script improvements
When a containerized agent node is set up by the quest tool (via the
setup script) we need to handle certification so the user doesn't need
to repeat the cert signing process at the beginning of each quest. After
the initial quests where we demonstrate the full process of a puppet
agent run, including pluginsync, we want to pre-run pluginsync as a node
is created so the pluginsync output on their first manually triggered
puppet agent run will not be overwhelmed by the pluginsync output.

Prior to this PR, a full puppet run is triggered during node creation to
create a CSR, then another is triggered after the cert is signed to
handle pluginsync. However, this solution can lead to problems, as the
second puppet run will also generate a catalog and apply it to the node.
If the user has written code with an error, there may be issues with
this initial puppet run that they do not see because it is completed in
the background.

This PR addresses this issue by separating the cert signing process and
the initial puppet run into two distinct actions. The ssl files are
generated on the master and copied with cp to an ssl directory mounted
to the containers. This allows the certs to be handled outside the
context of a puppet agent run. The initial puppet agent run is then
triggered with a bogus `--tags bogus` parameter. This ensures that no
user-specified resources are actually applied on the node.

Note, however, that it is still possible that this initial run will fail
if there is a catalog compilation error.