(LEARNVM-635) Add PuppetHack environment option to setup
Add an option to create a set of nodes to be used in 2017 PuppetHack
event. These will be available when running the setup script directly,
with the puppethack argument, but will not be exposed by the quest tool.

Because the output of the setup script will be shown when run directly,
this also redirects the output of the cert signing commands (which are
expected to fail in some circumstances) to /dev/null. This will keep
users from becoming unduly concerned about red text in the setup
script's output.