Correct logic  error in example "if" test
The if statement is not logically valid.

if $sinatra_server == 'thin' or 'mongrel'

Due to order of operations, == happens before "or". Therefore this test
always evaluates true, because 'mongrel' is a non-empty string.

Also fixes an error in a spec test where \1 was used in a place
that \2 should have been used.